AIC and predict svyolr modifications for Ordered Logit [revdep skip]
Temporarily compute regular AIC using provided svyolr outputs. This needs to be
re-examined to align with the dAIC computed in other svy functions. Also the predict
function is possible if the svyolr model output is coerced to polr class. Note the
svyolr function was heavily copied from polr as the author Thomas Lumley states