Add error message handler for svyolr with interaction
svyolr needs to invert the Hessian matrix to compute standard errors.
There is a good chance an interaction crosstab could result in unobserved
levels in the response variable giving a singular Hessian. Since the crosstab
cannot be created in this context, an error in intercepted and given a more
informative message. Unit tests also updated.