Improved error handling of svyolr
Improved the error message to be more empathetic and informative when
interaction tests fail for weighted ordered logit (svyolr regression).
Also throwing an informative error that Anova diagnostics cannot be
computed for svyolr class objects. Also protecting the call to
FitRegression in estimateRelativeImportance to isolate the previously
mentioned error message to occur when there are unobserved levels in
an interaction test