Refactor FitRegression to allow multiple calls
Refitting the data when outlier.proportion is set should call the same code twice instead of duplicating code for the second run. To facilitate this the subset argument of the lm, glm and survey functions will be used. This required a refactor to add a logical vector to the data