DS-2498 (#7)
* Introduce syvolr to replace polr for Ordered Logit [ci skip]

* AIC and predict svyolr modifications for Ordered Logit [revdep skip]

Temporarily compute regular AIC using provided svyolr outputs. This needs to be
re-examined to align with the dAIC computed in other svy functions. Also the predict
function is possible if the svyolr model output is coerced to polr class. Note the
svyolr function was heavily copied from polr as the author Thomas Lumley states

* Further adapt svyolr output to handle predict methods

svyolr output is very similar to polr output but missing two elements.
The model (data) and lp (linear predictor). Since the svyolr class was
coerced to polr to integrate into existing codebase, the model element is
required since it is used in the predict.Regression method. Commit adds
the model element to the svyolr list output.

* Add error message handler for svyolr with interaction

svyolr needs to invert the Hessian matrix to compute standard errors.
There is a good chance an interaction crosstab could result in unobserved
levels in the response variable giving a singular Hessian. Since the crosstab
cannot be created in this context, an error in intercepted and given a more
informative message. Unit tests also updated.

* Add dAIC calculation for svyolr

The survey design adjusted AIC is calculated for the svyolr models.
This is computed in the fitOrderedLogit function and follows the
theoretical paper of Lumley and Scott (see DS-2498 Jira task)

* Improved error handling of svyolr

Improved the error message to be more empathetic and informative when
interaction tests fail for weighted ordered logit (svyolr regression).
Also throwing an informative error that Anova diagnostics cannot be
computed for svyolr class objects. Also protecting the call to
FitRegression in estimateRelativeImportance to isolate the previously
mentioned error message to occur when there are unobserved levels in
an interaction test

* Improve documentation and add ... argument to svyolr [revdep skip]

* Add linear predictor to svyolr output for later use

The task in DS-2704 requires computation of residuals for ordered logit models. The sure::resids function is planned to be used for this purpose and requires the linear predictor to integrate into the polr implementation in sure. Consequently the linear predictor output is computed in the fitOrderedLogit function