Fix data in Regression output for Multiple Imputation
Data is saved in the Regression output list element in object\$model. This is used for the predict method to predict responses. In the missing = 'Multiple imputation' case this element was not updated to contain the entire original dataset and instead only had filtered data from the first iteration for multiple imputed data. This commit fixes this and amends the unit test to check the output.

Similar fix required for weights, the overall weights, not the processed
imputed data weights are saved. This is tested in flipRegressionTests