Further validate input and add formula [ci skip]
Add checks to further validate the input data to be stacked. Also
refactored slightly to rename the resulting data.frame to have generic
column names (Y and X1, ... Xn) where n is the number of predictors. The
variable labels are added as attributes to each column in the data
frame.

Code to construct the formula has been added created simply using the
basic default variable names above while allowing the labels to be used
for other outputs.