Add logic to use subset, weights and interaction
Added logic to ensure that the subset, weight and interaction arguments
are appropriately expanded to be compatible with the newly stacked
data.

Also added unit tests to ensure consistency between Regression outputs
on data that is already stacked and the new output that stacks the data
automatically.

A new dataset is added and documented to facilitate the tests described
above.