Give default formula argument [revdep skip]
Previously a formula was always required to determine the Regression
model. This is no longer the case since the formula can be derived when
the user inputs stackable data via the 'stacked.data.check' and
'unstacked.data' arguments. The default formula argument is set to be
'as.formula(NULL)' since the environment attribute of the formula is
used in the Regression call in the stackable case.

Throws an error when no formula is provided in the regular data context
and gives an informative error message. Unit tests also updated to
reflect this change.