Improve warnings and allow interaction in dummy adjustment
Improved some warnings and messages. A previously made warning in another development branch incorrectly described all variables in df, regardless of their structure, being converted to categorical when it should state only the ones that are converted from ordinal to categorical. Also added logic to remove dummy variables from the printing when a Regression with interaction is used.