DS-83 Ensure robust SE works with Dummy var adj
The robust standard errors rescales the variance estimates based off the
squared residuals and the influence values (hat values). If the dummy
variable adjustment is used in scenarios where there isn't enough missing
data observed for each predictor. The dummy variable adjustment will overfit
the missing values and give a perfect fit for the outcomes with missing predictors
This causes the hat values for these cases to be one and cause singularities
in the HCCM (Heteroscedastic Consistent Covariance Matrix). This commit will
catch cases that have unit hat values and adjust the HCCM calculation to use
hc1 instead of hc3. hc1 uses a degree of freedom adjustment and not
influence adjustment. This will allow robust standard errors to be computed