DS-2704 Update footer in outlier removal case [revdep skip]
Updating the footer wasn't correctly implemented to handle the case of
automated outlier removal. The previous method recalled the
SampleDescription function with new estimation and subset counts based off
the outlier removal settings. However, these counts are used in
SampleDescription to determine if there is missing data and caused
footer to display that there is missing data when there is none. This
commit fixes this by using a regex to intercept the sample description
but only the first base description part (all text up to the first
semi-colon). Later text contains missing data information and is not
touched. A new updated base description is substituted over it.