DS-2703 3 (#16)
* DS-2703 Copy attributes in weights and interaction

The Stack data option can handle filters, weights and interaction by
taking the usual filters, weights and interaction and stacking them
appropriately with a call to `rep`. Previously the attributes were not
copied in this process and metadata lost. This commit fixes this.

* DS-2703 Improve tests and check new warning

Warning has been improved in flipTransformations that is now checked in
this new commit. Also the unit tests for weights and filter attributes
have been improved.

* Update travis build order 
* DS-2703 Minor test update for warning msg

flipTransformsations@DS-2703 has the warning message updated to minimise
downstream conflicts. This commit aligns with that change

* Remove branch references in .travis.yml

Co-authored-by: chschan <chschan@users.noreply.github.com>