DS-2824 Generalize HCCM handling
Previously the commits in this branch only considered HCCM adjustments
in the scope of dummy varible adjustment. This has been relaxed and it
is applied more generally to catch completely influential observations.
The functions have been tidied to reduce duplication and messages
tweaked to be more informative to the user.