DS-2824 Ensure robust SE works with Dummy var adj (#17)
* DS-83 Ensure robust SE works with Dummy var adj

The robust standard errors rescales the variance estimates based off the
squared residuals and the influence values (hat values). If the dummy
variable adjustment is used in scenarios where there isn't enough missing
data observed for each predictor. The dummy variable adjustment will overfit
the missing values and give a perfect fit for the outcomes with missing predictors
This causes the hat values for these cases to be one and cause singularities
in the HCCM (Heteroscedastic Consistent Covariance Matrix). This commit will
catch cases that have unit hat values and adjust the HCCM calculation to use
hc1 instead of hc3. hc1 uses a degree of freedom adjustment and not
influence adjustment. This will allow robust standard errors to be computed

* DS-2824 Adjust only divisors in hc3 calc [revdep skip]

Previously the entire method was changed from hc3 (default) to hc1 when
there was a singular h (causing division by zero). This commit just
changes the adjusted residual computations to use the degree of freedom
adjusted residual estimates instead of the influence adjusted. There
will be no singularity and computation should proceed. However, the
covariance matrix itself might be unstable as expected and warning
thrown.

* DS-2824 Generalize HCCM handling

Previously the commits in this branch only considered HCCM adjustments
in the scope of dummy varible adjustment. This has been relaxed and it
is applied more generally to catch completely influential observations.
The functions have been tidied to reduce duplication and messages
tweaked to be more informative to the user.

* DS-2824 Remove references to dummy vars in docs [ci skip]

* fixup! DS-2824 Remove references to dummy vars in docs [revdep skip]