DS-2836 Utilize codeframe names for stacking
Use the codeframes for name acquisition in data stacking. Refactored the
code to allow stacking with the Q/Displayr codeframe or manual stacking
that doesnt have the codeframe information. In the latter, it will
require comma separation in variable names and throw an error if a
variable label uses commas (ambiguous separation). Unit tests have been
updated to accommodate both styles and 100% coverage achieved across the
stacking functions.