DS-2876: Check constant variables for Jaccard [revdep skip]
Previously check for binary variables too rigid where both 0 and 1 values are required in a variable. This commit relaxes this to allow constant variables (all zero or all ones). In the case of constant variables, a warning is thrown to the user alerting them to the constant. Unit tests added to check warning is thrown in those contexts. Also tidied up print method to not alert user to count data in Jaccard Coefficient output