DS-2876 Fix edge case error in setCor
The setCor function in psych checks if a correlation matrix is valid by
a simple check max(x) <= 1 & min(x) >= -1. The diagonal elements of the
computed weighted correlation matrix can be 1 within machine epsilon
tolerance. This commit forces the diagonals to be one. This is to be
tested in the Driver Analysis Standard R tests on Correlation.