DS-2919 Improve error handler for polr/svyolr
Error handler ran code assuming that svyolr had been used and .design
existed in the environment. This commit fixes this oversight and also
improves the error message if a user attempts to fit an Ordered Logit
when the outcome variable has two or fewer levels. In the former case, a
binary logit model is selected. In the latter the user is advised that
the outcome needs variation for a Regression to be computed. Unit tests
updated.