DS-2964 Allow stacking to handle codeframe switch
Current stacking code assumed the outcome names were in the codeframe
and the predictor names were in the secondarycodeframe. This commit
allows the reverse scenario to be handled as well. Also reworded some of
the warnings to be Q/Displayr agnostic. Errors are now thrown if there
is ambiguity (name clashes) in the outcome and predictor names.