DS-2964 Remove NETs during stacking
The code to identify and flag columns for removal during data stacking
has been updated. It will now identify and remove columns from the data
that consist of NETs that are not entirely comprised of unique variables
that are not observed elsewhere in the data. That is, the NET will be
removed if any variable in its composition is observed elsewhere. This
is guided by the information in the (secondary)codeframe as necessary.