DS-2964 Identify NETs with hidden and observed codes
Identify the NETs in the codeframe when stacking that are comprised of
both hidden and observed codes. These NETs are to be removed by the user
needs to be aware that they are being removed and the hidden code
information is gone with it and cannot be recovered without changing the
input data. A warning is given to advise them of this.