DS-2964 Remove NETs during stacking (#29)
* DS-2964 Remove NETs during stacking

The code to identify and flag columns for removal during data stacking
has been updated. It will now identify and remove columns from the data
that consist of NETs that are not entirely comprised of unique variables
that are not observed elsewhere in the data. That is, the NET will be
removed if any variable in its composition is observed elsewhere. This
is guided by the information in the (secondary)codeframe as necessary.

* DS-2964 Identify NETs with hidden and observed codes

Identify the NETs in the codeframe when stacking that are comprised of
both hidden and observed codes. These NETs are to be removed by the user
needs to be aware that they are being removed and the hidden code
information is gone with it and cannot be recovered without changing the
input data. A warning is given to advise them of this.