DS-2990 Throw informative error if aliased preds in RIA
Checks the design matrix for aliased predictors before any Shapley
regression or Relative Importance Analysis is conducted (it cannot
handle linearly dependent predictors). If linearly dependent predictors
are found then the linearly dependent combinations of variables are
collated and thrown in an informative message to the user that the
analysis cannot occur with that combination of variables.