DS-2990-2 Add no variation checks and update crosstab
Added checks to ensure predictors and outcome variables have variation
before being passed to estimateImportance for Shapley and RIA. Also
updated the crosstab interaction code to inspect the data and validate
it for Shapley and RIA before computation. The computation is skipped
for the split group if it doesnt pass the checks.