DS-2884 Fix Ordered Logit with disambiguated var. names (#33)
* DS-2990 Tidy up dummy adjusted cat preds in RIA

Dummy adjustment of categorical predictors is not supported in RIA but
categorical predictors are possible if they are not dummy adjusted. This
commit enables this to occur.

* DS-2990 Tidy up formula, names and labels

Simplifed the code that determines the formula for Jaccard output,
removing the formula with interaction since it is not required. Also
added names and labels elements to the return list since the same
characters are determined in multiple places.

* DS-2990 Throw informative error if aliased preds in RIA

Checks the design matrix for aliased predictors before any Shapley
regression or Relative Importance Analysis is conducted (it cannot
handle linearly dependent predictors). If linearly dependent predictors
are found then the linearly dependent combinations of variables are
collated and thrown in an informative message to the user that the
analysis cannot occur with that combination of variables.

* DS-2990-2 Add no variation checks and update crosstab

Added checks to ensure predictors and outcome variables have variation
before being passed to estimateImportance for Shapley and RIA. Also
updated the crosstab interaction code to inspect the data and validate
it for Shapley and RIA before computation. The computation is skipped
for the split group if it doesnt pass the checks.

* WIP ordered logit with multiple data sets; DS-2884

* Bump version, rm debugging from last commit; DS-2884

* Fix case where data set needs backticks; DS-2884

* Also, fix unit test and add additional one

Co-authored-by: jrwishart <justin.wishart@displayr.com>
Co-authored-by: flipDevTools <opensource@displayr.com>
Co-authored-by: Justin Yap <justinyapcc@gmail.com>