DS-2990 Shield RIA validation from data set refs
Formulae that use dataset refs such as `My data`$Variables$Y are not
variable references when used outside the environment of flipRegression.
It caused an issue in stats::alias that is fixed here by santizing the
formula and data, stripping out the dataset referencing and
disambiguating the variable names if necessary.