DS-2988 Prevent backticks from being stripped in alias checks
In RIA, checks are performed for aliased variables and the alias function
is used. It seems it strips off backticks from variable names in formulae
which can cause an error since the formula variables dont occur in the
supplied data. This is fixed with a call to make.names to remove the backticks
in a safe way before being used in alias.