DS-4096 Add code to validate Regression arguments
Some settings are not valid or supported in a call to Regression. e.g.
Multiple imputation with automated outlier removal. In these cases the
function should error early with an informative message. The code has
been improved to catch such combinations early and throw informative
messages through a helper function at the beginning of the call to
Regression.